Moved Composables and Model to individual files
