+++ app/src/main/kotlin/at/bitfire/davdroid/ui/AppSettingsActivity.kt
@@ -79,11 +79,11 @@ import at.bitfire.davdroid.util.broadcastReceiverFlow
 import at.bitfire.ical4android.TaskProvider
 import dagger.hilt.android.AndroidEntryPoint
 import dagger.hilt.android.lifecycle.HiltViewModel
+import javax.inject.Inject
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
-import javax.inject.Inject
 
 @AndroidEntryPoint
 class AppSettingsActivity: AppCompatActivity() {
@@ -541,7 +541,7 @@ class AppSettingsActivity: AppCompatActivity() {
         fun resetHints() {
             settings.remove(BatteryOptimizationsPage.Model.HINT_BATTERY_OPTIMIZATIONS)
             settings.remove(BatteryOptimizationsPage.Model.HINT_AUTOSTART_PERMISSION)
-            settings.remove(TasksActivity.Model.HINT_OPENTASKS_NOT_INSTALLED)
+            settings.remove(TasksModel.HINT_OPENTASKS_NOT_INSTALLED)
             settings.remove(OpenSourcePage.Model.SETTING_NEXT_DONATION_POPUP)
         }
