+++ app/src/main/kotlin/at/bitfire/davdroid/ui/intro/TasksIntroPage.kt
@@ -7,8 +7,8 @@ package at.bitfire.davdroid.ui.intro
 import android.app.Application
 import androidx.compose.runtime.Composable
 import at.bitfire.davdroid.settings.SettingsManager
-import at.bitfire.davdroid.ui.TasksActivity
 import at.bitfire.davdroid.ui.TasksCard
+import at.bitfire.davdroid.ui.TasksModel
 import at.bitfire.davdroid.util.TaskUtils
 import dagger.hilt.EntryPoint
 import dagger.hilt.InstallIn
@@ -26,7 +26,7 @@ class TasksIntroPage : IntroPage {
     override fun getShowPolicy(application: Application): IntroPage.ShowPolicy {
         val settingsManager = EntryPointAccessors.fromApplication(application, TasksIntroPageEntryPoint::class.java).settingsManager()
 
-        return if (TaskUtils.isAvailable(application) || settingsManager.getBooleanOrNull(TasksActivity.Model.HINT_OPENTASKS_NOT_INSTALLED) == false)
+        return if (TaskUtils.isAvailable(application) || settingsManager.getBooleanOrNull(TasksModel.HINT_OPENTASKS_NOT_INSTALLED) == false)
                 IntroPage.ShowPolicy.DONT_SHOW
             else
                 IntroPage.ShowPolicy.SHOW_ALWAYS
